Fix for playlist display in web app
If changing the playlist via the web app, the old playlist is displayed after hitting the "play" button which loads the a playlist.
This is because the playout_control.sh is called in the background from rfid_trigger_play.sh and so the web app doesn't wait for the successful execution of the play command and reloads the page immediately. So the inc.playerStatus get's the old playlist as the new playlist is not loaded yet (this is a matter of milliseconds).