fix: handling of FOLDER in playlistaddplay
The code in `inc.settingsFolderSpecific.sh` should read the value for 
`FOLDER` only from `Latest_Folder_Played` if it is not set by its parent 
script.
Otherwise this might return with no value at all for example on the 
first media played and with wrong settings if changing playlists.

fixes #514