Update ajax.loadMopidyStatus.php
The problem is, that mopidy.SERVICE (Spotify) is starting very quick. But while the service is already running, mopidy is establishing the SERVER connection to localhost:6600 in the background. This need 45-60 seconds, depending on how much playlists the user has.

Only when the server is running, you can give commands (volume, starting songs from spotify, ...)

By the way, in the information tab you can still see, the mopidy server status. But this is only the status of the service.
I added seperate informations about service and server.