Wait until server is running (#1026)
* Update startup-scripts.sh

Adding a while-loop for checking the mopidy/MPD server status.
The startup-script continues when server is up and running.
So, now the startup-sound indicates that the box and all packages are ready to use.

* Update ajax.loadMopidyStatus.php

in the information tab you can still see, the mopidy server status. But this is only the status of the service.
I added seperate informations about service and server.