Fix incorrect ownership of ~/.config/phoniebox
The `chown` command in line 61 changes the ownership of the gpio_settings.ini to the correct user. Because the script is run with sudo command root is the of ~/.config/phoniebox  which leads to an error in gpio_control.py. The altered `chown` command leads to the folder and the the .ini to be owned by the correct user.