Example for new Evdev config structure
This is tighly modeled after gpio.
In contrast to GPIO, this is a list of devices containing
each one or more input/output devices.
Eg a Joystick has keys but potentially also rumble or leds.

Currently only very simply input devices (buttons) are supported.
This structer should enable to extend this easily.