Adapt evdev code to new config
Now the config for evdev is akin to GPIO.
The big difference is that multiple devices with each
multiple input/output devices are supported.

Note that the backend is still the old event device listener backend.
Thus to support more features, the backend would need to be
quite drastically overhauled.